LPD-75968 portal-workflow-metrics-service: Move workflow metrics reindexers from portal-workflow-kaleo-metrics-integration
https://liferay.atlassian.net/browse/LPD-75968
