+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-service/src/main/java/com/liferay/portal/workflow/metrics/internal/search/index/reindexer/TaskWorkflowMetricsReindexer.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.workflow.kaleo.metrics.integration.internal.search.index.reindexer;
+package com.liferay.portal.workflow.metrics.internal.search.index.reindexer;
 
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.Property;
@@ -12,10 +12,12 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.workflow.kaleo.metrics.integration.helper.IndexerHelper;
 import com.liferay.portal.workflow.kaleo.model.KaleoDefinitionVersion;
 import com.liferay.portal.workflow.kaleo.model.KaleoInstance;
+import com.liferay.portal.workflow.kaleo.model.KaleoTaskInstanceToken;
 import com.liferay.portal.workflow.kaleo.service.KaleoDefinitionVersionLocalService;
 import com.liferay.portal.workflow.kaleo.service.KaleoInstanceLocalService;
+import com.liferay.portal.workflow.kaleo.service.KaleoTaskInstanceTokenLocalService;
 import com.liferay.portal.workflow.metrics.search.background.task.WorkflowMetricsReindexStatusMessageSender;
-import com.liferay.portal.workflow.metrics.search.index.InstanceWorkflowMetricsIndexer;
+import com.liferay.portal.workflow.metrics.search.index.TaskWorkflowMetricsIndexer;
 import com.liferay.portal.workflow.metrics.search.index.reindexer.WorkflowMetricsReindexer;
 
 import java.util.Objects;
@@ -28,18 +30,17 @@ import org.osgi.service.component.annotations.Reference;
  * @author Rafael Praxedes
  */
 @Component(service = WorkflowMetricsReindexer.class)
-public class InstanceWorkflowMetricsReindexer
-	implements WorkflowMetricsReindexer {
+public class TaskWorkflowMetricsReindexer implements WorkflowMetricsReindexer {
 
 	@Override
 	public String getKey() {
-		return "instance";
+		return "task";
 	}
 
 	@Override
 	public void reindex(long companyId) throws PortalException {
 		ActionableDynamicQuery actionableDynamicQuery =
-			_kaleoInstanceLocalService.getActionableDynamicQuery();
+			_kaleoTaskInstanceTokenLocalService.getActionableDynamicQuery();
 
 		actionableDynamicQuery.setAddCriteriaMethod(
 			dynamicQuery -> {
@@ -54,34 +55,32 @@ public class InstanceWorkflowMetricsReindexer
 		AtomicInteger atomicCounter = new AtomicInteger(0);
 
 		actionableDynamicQuery.setPerformActionMethod(
-			(KaleoInstance kaleoInstance) -> {
+			(KaleoTaskInstanceToken kaleoTaskInstanceToken) -> {
 				KaleoDefinitionVersion kaleoDefinitionVersion =
 					_kaleoDefinitionVersionLocalService.
 						fetchKaleoDefinitionVersion(
-							kaleoInstance.getKaleoDefinitionVersionId());
+							kaleoTaskInstanceToken.
+								getKaleoDefinitionVersionId());
 
 				if (Objects.isNull(kaleoDefinitionVersion)) {
 					return;
 				}
 
-				_instanceWorkflowMetricsIndexer.addInstance(
-					_indexerHelper.createAssetTitleLocalizationMap(
-						kaleoInstance.getClassName(),
-						kaleoInstance.getClassPK(), kaleoInstance.getGroupId()),
-					_indexerHelper.createAssetTypeLocalizationMap(
-						kaleoInstance.getClassName(),
-						kaleoInstance.getGroupId()),
-					kaleoInstance.getClassName(), kaleoInstance.getClassPK(),
-					companyId, kaleoInstance.getCompletionDate(),
-					kaleoInstance.getCreateDate(),
-					kaleoInstance.getKaleoInstanceId(),
-					kaleoInstance.getModifiedDate(),
-					kaleoInstance.getKaleoDefinitionId(),
-					kaleoDefinitionVersion.getVersion(),
-					kaleoInstance.getUserId(), kaleoInstance.getUserName());
+				KaleoInstance kaleoInstance =
+					_kaleoInstanceLocalService.fetchKaleoInstance(
+						kaleoTaskInstanceToken.getKaleoInstanceId());
+
+				if (Objects.isNull(kaleoInstance)) {
+					return;
+				}
+
+				_taskWorkflowMetricsIndexer.addTask(
+					_indexerHelper.createAddTaskRequest(
+						kaleoInstance, kaleoTaskInstanceToken,
+						kaleoDefinitionVersion.getVersion()));
 
 				_workflowMetricsReindexStatusMessageSender.sendStatusMessage(
-					atomicCounter.incrementAndGet(), total, "instance");
+					atomicCounter.incrementAndGet(), total, "task");
 			});
 
 		actionableDynamicQuery.performActions();
@@ -91,14 +90,18 @@ public class InstanceWorkflowMetricsReindexer
 	private IndexerHelper _indexerHelper;
 
 	@Reference
-	private InstanceWorkflowMetricsIndexer _instanceWorkflowMetricsIndexer;
-
-	@Reference
 	private KaleoDefinitionVersionLocalService
 		_kaleoDefinitionVersionLocalService;
 
 	@Reference
 	private KaleoInstanceLocalService _kaleoInstanceLocalService;
+
+	@Reference
+	private KaleoTaskInstanceTokenLocalService
+		_kaleoTaskInstanceTokenLocalService;
+
+	@Reference
+	private TaskWorkflowMetricsIndexer _taskWorkflowMetricsIndexer;
 
 	@Reference
 	private WorkflowMetricsReindexStatusMessageSender
